fix(ci): extract merge-attempt + attention primitives to clear first-party line-limit guardrail
python-coverage-shards (8) failed
test_first_party_code_files_stay_under_line_limit because
merge_loop.py (1508) and lifecycle.py (1507) exceeded the 1500-line
limit after the #661/#663 attention-clear work.

Extract two cohesive groups into focused modules wired back through
imports and mixins.py (behavior unchanged):
- merge_attempts.py: _BITBUCKET_IN_FLIGHT_MERGE_REASON_CODES,
  _MergeAttemptOutcome, _MergeAttemptResult,
  _record_empty_effective_merge_methods_blocker, _attempt_merge_method
  (from merge_loop)
- merge_attention.py: _set_workspace_attention,
  _set_workspace_attention_with_merge_block_marker,
  _clear_stale_merge_attention, _clear_merge_block_attention_durably,
  _clear_workspace_attention, _persist_merge_block_attention_durably
  (from lifecycle)

Test import of _MergeAttemptOutcome/_MergeAttemptResult follows the
symbols to merge_attempts. merge_loop.py -> 1151, lifecycle.py -> 1182.
